+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 This is the changelog of our "JTL WooCommerce Connector".
 
 ## Unreleased
+- CO-2494 - disable cache
 
 ## 1.41.1 _2023-11-28_
 - CO-2466 - fix undefined method in manufacturer model